+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ###v0.1.5
 - Added `expose_data` method to `MustacheView`
 - Added `Controller_Mustache` class
+- Added an example of how to expose ORM models
 
 ###v0.1
 First version